refactor(workstation): extract history surface from inkRuntime (#890 phase 5a.5)
Fifth and final surface-family extraction. The history surface is the
workstation's home screen — commit list with optional graph lanes,
synthetic pending-commit row, and server-side filter indicator. Pulls
the family of five paired functions out of inkRuntime in one cut since
they only call each other:

  - renderHistoryPanel        (the surface itself, exported)
  - renderLaneSegmentSpans    (#791 stage 2 lane-color rendering)
  - renderCommitHistoryRow    (per-commit row with theme-aware spans)
  - renderPendingCommitRow    ((+) new commit affordance)
  - formatHistoryFetchArgs    (path: / author: filter indicator)

All five now live in src/workstation/surfaces/history/index.ts (295 LOC).
Only `renderHistoryPanel` is exported — the rest are private to the
module, mirroring how they lived as inkRuntime locals.

inkRuntime.ts: 4,305 -> 4,040 LOC (-265 this PR; -1,999 cumulative since
5a started, -33% off the original 6,039). 9 imports pruned post-extraction:
  GitLogCommitRow (commands/log/data),
  formatInkRefLabels, getVisibleLogInkHistory (chrome/historyRows — whole import gone),
  substituteGraphChars (chrome/graphChars — whole import gone),
  LaneSegment, getLaneColor (chrome/graphLanes — whole import gone),
  formatLogInkHistoryEmpty (chrome/surfaceStates),
  cellWidth (chrome/text),
  LogInkHistoryFetchArgs (commands/log/inkViewModel).

ALL 12 main surfaces (plus conflicts) are now extracted:
  ✓ reflog, bisect, stash, worktrees     (5a.1)
  ✓ compose, branches, tags, pullRequest (5a.2)
  ✓ status + conflicts                   (5a.3)
  ✓ diff                                 (5a.4)
  ✓ history                              (this PR)

What's left in inkRuntime.ts at 4,040 LOC:
  - LogInkApp React component (~1,940 LOC, phase 5b target)
  - Chrome render helpers (header, sidebar, footer, overlays, palette,
    command palette) (~1,200 LOC, phase 5a.6 target)
  - Detail panel + inspector family (~600 LOC, phase 5a.6 target)
  - Preview panels (~150 LOC)
  - Imports + small orchestration helpers + startInkInteractiveLog
    + loadInkRuntime (~150 LOC)

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.4). No behavior change — history
renders byte-identically before/after.
